Refactor app structure for improved routing and efficiency
- Reorganized app routing to be more intuitive and Next.js-friendly.
- Enhanced authentication mechanisms for better security and user experience.
- Optimized navigation components for smoother transitions and clearer user flow.
- Addressed feedback and incorporated best practices for software architecture.
